Shrink ExportedReportPage by creating a new component
ExportedReportPage was managing the title AND the metadata table AND the
combined data table (for MCPAR) AND calling out to child components to
display the actual report data. This was clearly too much.

By pulling the metadata table out into a component, we
 * Reduce the size of the page file by 30%
 * Remove the need for several functions which returned JSX.Element
 * Therefore also remove the need to any ESLint config changes

The remainder of this commit's changes are typo fixes, formatting
fixes, or direct consequences of the new ExportedReportMetadataTable.

The larger question of "what is the best way to handle report-specific
verbiage across the project?" remains unresolved by this commit. I would
like to try out some different approaches, but they are out of scope
for this story.

Related work items: MDCT-2227